requirements.txt
transformers
torch
pillow
tk
